apps/ingest/requirements.txt
pdfplumber==0.11.0
pdfminer.six==20231228
beautifulsoup4==4.12.3
requests==2.32.3
PyYAML==6.0.2
typesense==0.21.0
